requirements.txt
python-dotenv
flask 
google-generativeai 
nltk 
numpy 
spacy 
gtts 
moviepy 
pydub 
pillow